requirements.txt
streamlit
openai-whisper
scenedetect[opencv]
opencv-python-headless
numpy